# swagger_mainnet.yaml
openapi: 3.0.1

info:
  title: Bitmap
  description: "Bitmap Inscription Service"
  version: 0.0.1

servers:
  - url: "http://localhost:8081/api"
  - url: "https://file-inscription-backend.onrender.com/api"

tags:
  - name: Status
  - name: Inscribe
  - name: Estimate

paths:
  /status/price:
    get:
      summary: Get Current bitcoin price
      tags:
        - Status
      responses:
        "200":
          description: Current Bitcoin Price
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Get Bitcoin Price Failed!
          content:
            application/json:
              schema:
                type: string
  /status/avgFeeRate:
    get:
      summary: Get last 15 blocks average feeRate
      tags:
        - Status
      responses:
        "200":
          description: Last 15 blocks average feeRate
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Get average feeRate failed!
          content:
            application/json:
              schema:
                type: string
  /status/recommendFeeRate:
    get:
      summary: Get recommended block feeRate
      tags:
        - Status
      responses:
        "200":
          description: Recommended block feeRate Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Recommended block feeRate Failed!
          content:
            application/json:
              schema:
                type: string
  /wallet/utxo-split:
    get:
      summary: Split large UTXO of Admin wallet
      tags:
        - Status
      responses:
        "200":
          description: Split large UTXO of Admin wallet Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Split large UTXO of Admin wallet Failed!
          content:
            application/json:
              schema:
                type: string

  /inscribe/text:
    post:
      summary: Create new Text Inscription
      tags:
        - Inscribe
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TextInscription"
      responses:
        "200":
          description: Text Inscription Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Text Inscription Failed!
          content:
            application/json:
              schema:
                type: string
  /inscribe/file:
    post:
      summary: Create new file inscription.
      tags:
        - Inscribe
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileInscription"
      responses:
        "200":
          description: File Inscription Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: File Inscription Failed!
          content:
            application/json:
              schema:
                type: string
  /inscribe/delegate:
    post:
      summary: Create new Delegate Inscription
      tags:
        - Inscribe
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DelegateInscription"
      responses:
        "200":
          description: Delegate Inscription Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Delegate Inscription Failed!
          content:
            application/json:
              schema:
                type: string
  /sendOrdinal/getSendingOrdinalBtcPsbt:
    post:
      summary: Get new Sending Ordinals Psbt
      tags:
        - Inscribe
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/SendingOrdialBtcPsbt"
      responses:
        "200":
          description: Get new Sending Ordinals Psbt Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Get new Sending Ordinals Psbt Failed!
          content:
            application/json:
              schema:
                type: string

  /estimate/text:
    post:
      summary: Estimate Text Inscription
      tags:
        - Estimate
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/TextInscription"
      responses:
        "200":
          description: Estimate Text Inscription Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Estimate Text Inscription Failed!
          content:
            application/json:
              schema:
                type: string
  /estimate/file:
    post:
      summary: Estimate file inscription
      tags:
        - Estimate
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileInscription"
      responses:
        "200":
          description: Estimate File Inscription Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Estimate File Inscription Failed!
          content:
            application/json:
              schema:
                type: string
  /estimate/delegate:
    post:
      summary: Estimate Delegate Inscription
      tags:
        - Estimate
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DelegateInscription"
      responses:
        "200":
          description: Estimate Delegate Inscription Success!
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Estimate Delegate Inscription Failed!
          content:
            application/json:
              schema:
                type: string

components:
  schemas:
    TextInscription:
      type: object
      properties:
        receiveAddress:
          type: string
          example: bc1pjr267uqlj79zqmv5m9vftvjp6m8fycwm8mq63497tyudq3kp3a7qh49jue
        contents:
          type: array
          items:
            type: string
            example: 350.364807.bitmap
        feeRate:
          type: number
          example: 30
        padding:
          type: number
          example: 546
        parentId:
          type: string
          example: 39adf8cde1d5a8906e362a1fb95412f5e4e0c8715422495163eb014679bcd901i0
        metadata:
          type: string
          example: '{"type":"dog","size":"medium"}'
        metaprotocol:
          type: string
          example: bitmap.community
        reinscriptionId:
          type: string
          example: 9566bc0e79ca905bf01db55bf154acb3a04e00e8bb2e9d4903639c5badf66d42i0
      required:
        - receiveAddress
        - contents
        - feeRate
        - padding
    FileInscription:
      type: object
      properties:
        receiveAddress:
          type: string
          example: bc1pjr267uqlj79zqmv5m9vftvjp6m8fycwm8mq63497tyudq3kp3a7qh49jue
        feeRate:
          type: number
          example: 30
        files:
          type: array
          items:
            type: string
            format: binary
        padding:
          type: string
          example: 546
        parentId:
          type: string
          example: 39adf8cde1d5a8906e362a1fb95412f5e4e0c8715422495163eb014679bcd901i0
        metadata:
          type: string
          example: '{"type":"dog","size":"medium"}'
        metaprotocol:
          type: string
          example: bitmap.community
        reinscriptionId:
          type: string
          example: 9566bc0e79ca905bf01db55bf154acb3a04e00e8bb2e9d4903639c5badf66d42i0
      required:
        - receiveAddress
        - feeRate
        - files
        - padding
    DelegateInscription:
      type: object
      properties:
        receiveAddress:
          type: string
          example: bc1pjr267uqlj79zqmv5m9vftvjp6m8fycwm8mq63497tyudq3kp3a7qh49jue
        delegateId:
          type: array
          items:
            type: string
            example: 9566bc0e79ca905bf01db55bf154acb3a04e00e8bb2e9d4903639c5badf66d42i0
        feeRate:
          type: number
          example: 30
        padding:
          type: number
          example: 546
        parentId:
          type: string
          example: 39adf8cde1d5a8906e362a1fb95412f5e4e0c8715422495163eb014679bcd901i0
        metadata:
          type: string
          example: '{"type":"dog","size":"medium"}'
        metaprotocol:
          type: string
          example: bitmap.community
        reinscriptionId:
          type: string
          example: 9566bc0e79ca905bf01db55bf154acb3a04e00e8bb2e9d4903639c5badf66d42i0
      required:
        - receiveAddress
        - delegateId
        - feeRate
        - padding
    SendingOrdialBtcPsbt:
      type: object
      properties:
        receiveAddress:
          type: string
          example: bc1pjr267uqlj79zqmv5m9vftvjp6m8fycwm8mq63497tyudq3kp3a7qh49jue
        feeRate:
          type: number
          example: 30
        parentId:
          type: string
          example: 39adf8cde1d5a8906e362a1fb95412f5e4e0c8715422495163eb014679bcd901i0
        reinscriptionId:
          type: string
          example: 9566bc0e79ca905bf01db55bf154acb3a04e00e8bb2e9d4903639c5badf66d42i0
        btcAmount:
          type: number
          example: 100000
        publicKey:
          type: string
          example: 03cd6f1283151e05d2c0a2ad6faa38c43e807d7039a68356df950c489957e4c754
      required:
        - receiveAddress
        - feeRate
        - btcAmount
        - publicKey